Tests/reorganize: move puny encoding test to own file
